# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bokdy API
  version: 0.1.0
  description: |
    Bokdy foundation API (Identity + Organization + CRM + Catalog + Scheduling + Pricing
    + Reservation + Booking + Billing + Payment).
    Locale is taken from Accept-Language (missing → vi). Read models return a resolved name;
    owner edit payloads use name_en + name_vi.
    Auth: send X-Client (player|owner|admin) on register, login, refresh, logout.
    Timestamps are UTC RFC3339. Rate limit: 429 TOO_MANY_REQUESTS.
    Error envelope: { code, message, details? }. Codes are UPPERCASE; see packages/config/error-codes.json.
servers:
  - url: http://localhost:8088
    description: Local
...
paths:
...
  /api/v1/organizations:
    get:
      tags: [Organizations]
      summary: List my organizations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
    post:
      tags: [Organizations]
      summary: Create organization
      description: Creates tenant, organization, default business unit (`code=default`), and owner staff with `org_owner`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/XOrganizationIdOptional"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Organization"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Organizations]
      summary: Update organization
      description: Owner or system admin. Does not change status.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/XOrganizationIdOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /api/v1/admin/organizations:
    get:
      tags: [Admin]
      summary: List all organizations
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Filter by code or name
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended, archived]
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: province_id
          in: query
          schema: { type: string, format: uuid }
          description: Filter orgs with at least one branch in this province (current_v2 address)
      responses:
        "200":
          description: Organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrganization"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/admin/organizations/{id}:
    get:
      tags: [Admin]
      summary: Get organization
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganization"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/admin/organizations/{id}/activate:
    post:
      tags: [Admin]
      summary: Activate organization
      description: |
        UC-ORG-003 workaround (no subscription): tenant `trial` → `active`, org `inactive` → `active`.
        Does not unsuspend. Already active is idempotent and does not emit a duplicate event.
        Emits OrganizationActivated when status changes.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Organization activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganization"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/organizations/{id}/suspend:
    post:
      tags: [Admin]
      summary: Suspend organization
      description: |
        Org `active` → `suspended` and tenant `trial|active` → `suspended`. Reason is required
        and stored on OrganizationSuspended only. Marketplace hides the org; staff membership
        routes and new holds/payments are rejected.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendOrganizationRequest"
      responses:
        "200":
          description: Organization suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganization"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/admin/organizations/{id}/restore:
    post:
      tags: [Admin]
      summary: Restore organization
      description: Suspended org and tenant → `active`. Emits OrganizationRestored.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Organization restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganization"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    XOrganizationIdOptional:
      name: X-Organization-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: If present on nested `/organizations/{id}/...` routes, must match path id.
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        default: vi
        example: vi
      description: BCP-47. First supported tag (vi, en) wins. Missing or unknown → vi.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Malformed JSON body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: BAD_REQUEST
            message: invalid json body
    ValidationFailed:
      description: Field validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: VALIDATION
            message: invalid request
            details:
              - field: email
                code: REQUIRED
                message: email is required
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - VALIDATION
            - BAD_REQUEST
            - TOO_MANY_REQUESTS
            - INTERNAL
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code:
          type: string
          enum:
            - REQUIRED
            - EMAIL_INVALID
            - TOO_SHORT
            - TOO_LONG
            - UUID_INVALID
            - ONE_OF
            - INVALID
            - PASSWORD_POLICY
        message: { type: string }
...
    Organization:
      type: object
      required: [id, public_id, tenant_id, code, name, status]
      properties:
        id: { type: string, format: uuid }
        public_id: { type: string, minLength: 26, maxLength: 26 }
        tenant_id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string, description: Resolved from Accept-Language }
        name_en: { type: string }
        name_vi: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        status: { type: string, enum: [active, inactive, suspended, archived] }
        tenant_status: { type: string, enum: [trial, active, suspended, canceled], description: Present on admin organization reads }
    AdminOrganization:
      allOf:
        - $ref: "#/components/schemas/Organization"
        - type: object
          required: [tenant_status, branch_count]
          properties:
            tenant_status: { type: string, enum: [trial, active, suspended, canceled] }
            branch_count: { type: integer, minimum: 0, description: Non-archived branch (location) count }
    SuspendOrganizationRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 1 }
    CreateOrganizationRequest:
      type: object
      description: At least one of name, name_en, or name_vi is required.
      properties:
        name: { type: string, description: Alias for name_vi when name_vi is omitted }
        name_en: { type: string }
        name_vi: { type: string }
        code: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
    UpdateOrganizationRequest:
      type: object
      properties:
        name_en: { type: string }
        name_vi: { type: string }
        code: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
...